Remove old sql source and rename the new one
