--- posthog/temporal/data_imports/pipelines/sql_database_v2/settings.py
@@ -1,1 +1,0 @@
-DEFAULT_CHUNK_SIZE = 10_000
